Added view dropdown, appearance changes
I added simple tree viewing options on the tool strip - expand and collapse. I got rid of that extra space on the bottom of the form, which also made its window fit on a 1366×768 screen, typical for laptops. I also changed Help to Links so that it's more obvious that it'll be opened in a browser.